% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlastrack_background}
\alias{atlastrack_background}
\title{AtlasTrack 2D background polygons}
\format{
A list with three \code{sf} objects:
\describe{
\item{gm}{Grey matter / cortical background polygons.}
\item{wm}{White matter cut-out polygons.}
\item{vent}{Ventricle polygons.}
}
}
\usage{
atlastrack_background
}
\description{
A list containing ggseg-style background layers for plotting AtlasTrack.
}
\keyword{datasets}
