% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atlastrack_tract_names}
\alias{atlastrack_tract_names}
\title{List available AtlasTrack tract names}
\usage{
atlastrack_tract_names()
}
\value{
A data frame containing AtlasTrack tract abbreviations and full names.
}
\description{
List available AtlasTrack tract names
}
