% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atlastrack_views_available}
\alias{atlastrack_views_available}
\title{Show available AtlasTrack views}
\usage{
atlastrack_views_available()
}
\value{
A data frame of available view sets and display views.
}
\description{
Show available AtlasTrack views
}
